had error in the deleting comments --> error was used an id for the buttons and i added an event lister thinking it will work knowing fully well id's target one element so i change it to class and used the event bubbling in javascript ..  now it working  can deletecomments without reloading the page.

+++ frontend/static/apijs/comment.js
@@ -16,7 +16,10 @@ class Comment extends Authorization{
     
      let respData = await resp.json()
 
-     return  respData
+    //  return  respData
+    
+        comment.displayAllComment(respData)
+
     }
 
 //     blog: 17
@@ -34,12 +37,12 @@ class Comment extends Authorization{
 
         this.noOfComments.innerText =`${data.length} Comments`;
         
-        
+        this.commentContainer.innerHTML = '';
         data.forEach(element=>{
             //this blgCommentBtn was sperated beacuse of this error  ${} javascript was complainng that i have to many nested ${}
             blogCommentBtn = `
-            <button class="bg-danger btn" id="deleteCommentBtn" data-commentID="${element.id}"  style="color: white;"">Delete</button>
-          <button class="bg-success btn" id="updateCommentBtn data-commentID="${element.id}" style="color: white;"  >Update</button>
+            <button class="bg-danger btn deleteCommentBtn" data-commentID="${element.id}"  style="color: white;">Delete</button>
+          <button class="bg-success btn updateCommentBtn " data-commentID="${element.id}" style="color: white;"  >Update</button>
     
           `
         
@@ -64,9 +67,7 @@ class Comment extends Authorization{
                        ''
                 :  //else dont show any thing 
                 ''
-
-
-                }
+             }
                 
           
             </li>
@@ -79,40 +80,41 @@ class Comment extends Authorization{
 
 
 
-comment = new Comment('#commentListContainer')
+comment = new Comment('#commentListContainer');
 
 comment.getAllComment(`/api/blog/comment/${blogId}/`)
 .then(data=>{
-    comment.displayAllComment(data)
-
-
     // since it an async code we need to do this in the promise
     // or else "deleteCommentBtn" -- will return false because it ran before the api call finished
-    let deleteCommentBtn = document.querySelector('#deleteCommentBtn')
-
+    // let deleteCommentBtn = document.querySelector('#deleteCommentBtn')
+    let commentListContainer = document.querySelector('#commentListContainer')
     // console.log(deleteCommentBtn)
-    deleteCommentBtn.addEventListener('click',e=>{
+    commentListContainer.addEventListener('click',e=>{
         e.preventDefault()
      
- 
+        if(e.target.className.includes('deleteCommentBtn')){
+            fetch(`/api/blog/comment/${e.target.dataset.commentid}/`, {
+                method: 'DELETE',
+                headers:{
+                    Authorization:`token ${JSON.parse(sessionStorage.getItem('userToken'))}`
+                }
+            })
+              .then(data=>data.status)
+              .then(data=>{
+                  if(data==204){
+                        comment.getAllComment(`/api/blog/comment/${blogId}/`)
+    
+                  }
+              })
+        }
 
-        fetch(`/api/blog/comment/${e.target.dataset.commentid}/`, {
-            method: 'DELETE',
-            headers:{
-                Authorization:`token ${JSON.parse(sessionStorage.getItem('userToken'))}`
-            }
-          })
-          .then(data=>data.status)
-          .then(data=>{
-              if(data==204){
-                  location.reload()
-              }
-          })
-   
+    
+    
     })
 
-})
 
+})
+//end of getAllcomment promise
